INCOMPLETE: Can't figure out right now.
It's using the reduce function to concatenate strings according to the 2nd parameter in the function, which could be any number. I'm not sure if I can use a rest parameter inside a reduce method. 
Additionally, when I tried to concatenate 'prev' and 'next' in the reduce method, I hjad issues.